before remove filters from indexes

+++ src/Controller/CompanyController.php
@@ -3,10 +3,10 @@
 namespace App\Controller;
 
 use App\Entity\Company;
+use App\Enum\UserRoleEnum;
 use App\Repository\CompanyRepository;
 use App\Repository\UserRepository;
 use Exception;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -21,8 +21,8 @@ class CompanyController extends BaseController
 
     public function __construct(
         ParameterBagInterface $params,
-        CompanyRepository $companyRepository,
-        UserRepository $userRepository
+        CompanyRepository     $companyRepository,
+        UserRepository        $userRepository
     )
     {
         $this->params = $params;
@@ -34,13 +34,15 @@ class CompanyController extends BaseController
     #[Route('/companies', methods: ['GET'])]
     public function index(Request $request): Response
     {
+        $data = json_decode($request->getContent(), true);
+        if (json_last_error() !== JSON_ERROR_NONE) {
+            return $this->json(['error' => 'Invalid JSON'], Response::HTTP_BAD_REQUEST);
+        }
         $filters = [
-            'pageNumber' => $request->query->getInt('page_number', 1),
-            'pageSize' => $request->query->getInt('page_size', 12),
+            'pageNumber' => $data['page_number'] ?? 1,
+            'pageSize' => $data['page_number'] ?? 12
         ];
-
         $companies = $this->companyRepository->index($filters);
-
         return $this->json($companies);
     }
 
@@ -48,25 +50,29 @@ class CompanyController extends BaseController
     public function show(int $id): Response
     {
         $company = $this->companyRepository->findById($id);
-
         if (!$company) {
             return $this->json(['error' => 'Company not found'], Response::HTTP_NOT_FOUND);
         }
-
         return $this->json($company);
     }
 
     #[Route('/companies/{id}/users', methods: ['GET'])]
-    public function getUsers(int $id): Response
+    public function getUsers(int $id, Request $request): Response
     {
+        $requestUser = $this->getUserFromJWT($request);
+        if (!$this->checkAccess($requestUser,
+            [UserRoleEnum::ROLE_SUPER_ADMIN, UserRoleEnum::ROLE_COMPANY_ADMIN])) {
+            return $this->responseForbidden();
+        }
+        if ($requestUser->isCompanyAdmin() and
+            $requestUser->getCompany()->getId() != $id) {
+            return $this->responseForbidden();
+        }
         $company = $this->companyRepository->findById($id);
-
         if (!$company) {
             return $this->json(['error' => 'Company not found'], Response::HTTP_NOT_FOUND);
         }
-
         $users = $this->companyRepository->getUsers($company);
-
         return $this->json($users);
     }
 
@@ -76,9 +82,16 @@ class CompanyController extends BaseController
     #[Route('/companies', methods: ['POST'])]
     public function create(Request $request, ValidatorInterface $validator): Response
     {
+        $requestUser = $this->getUserFromJWT($request);
+        if (!$requestUser->isSuperAdmin()) {
+            return $this->responseForbidden();
+        }
+        $data = json_decode($request->getContent(), true);
+        if (json_last_error() !== JSON_ERROR_NONE) {
+            return $this->json(['error' => 'Invalid JSON'], Response::HTTP_BAD_REQUEST);
+        }
         $company = new Company();
-        $company->setName($request->get('name'));
-
+        $company->setName($data['name']);
         $errors = $validator->validate($company);
         if (count($errors) > 0) {
             $errorMessages = [];
@@ -87,7 +100,6 @@ class CompanyController extends BaseController
             }
             return $this->json(['errors' => $errorMessages], Response::HTTP_BAD_REQUEST);
         }
-
         try {
             $this->companyRepository->store($company);
             return $this->json(['message' => 'Company created successfully'], Response::HTTP_CREATED);
@@ -99,14 +111,19 @@ class CompanyController extends BaseController
     #[Route('/companies/{id}', methods: ['PUT'])]
     public function update(int $id, Request $request, ValidatorInterface $validator): Response
     {
+        $requestUser = $this->getUserFromJWT($request);
+        if (!$requestUser->isSuperAdmin()) {
+            return $this->responseForbidden();
+        }
         $company = $this->companyRepository->findById($id);
-
         if (!$company) {
             return $this->json(['error' => 'Company not found'], Response::HTTP_NOT_FOUND);
         }
-
-        $company->setName($request->get('name'));
-
+        $data = json_decode($request->getContent(), true);
+        if (json_last_error() !== JSON_ERROR_NONE) {
+            return $this->json(['error' => 'Invalid JSON'], Response::HTTP_BAD_REQUEST);
+        }
+        $company->setName($data['name']);
         $errors = $validator->validate($company);
         if (count($errors) > 0) {
             $errorMessages = [];
@@ -115,7 +132,6 @@ class CompanyController extends BaseController
             }
             return $this->json(['errors' => $errorMessages], Response::HTTP_BAD_REQUEST);
         }
-
         try {
             $this->companyRepository->update($company);
             return $this->json(['message' => 'Company updated successfully']);
@@ -125,14 +141,16 @@ class CompanyController extends BaseController
     }
 
     #[Route('/companies/{id}', methods: ['DELETE'])]
-    public function delete(int $id): Response
+    public function delete(int $id, Request $request): Response
     {
+        $requestUser = $this->getUserFromJWT($request);
+        if (!$requestUser->isSuperAdmin()) {
+            return $this->responseForbidden();
+        }
         $company = $this->companyRepository->findById($id);
-
         if (!$company) {
             return $this->json(['error' => 'Company not found'], Response::HTTP_NOT_FOUND);
         }
-
         try {
             $this->companyRepository->delete($company);
             return $this->json(['message' => 'Company deleted successfully']);
